Move TemplateExpressionError to common.errros
