+++ pyomo/core/expr/template_expr.py
@@ -16,8 +16,8 @@ import sys
 import builtins
 
 from pyomo.common.backports import nullcontext
+from pyomo.common.errors import TemplateExpressionError
 from pyomo.core.expr.base import ExpressionBase, ExpressionArgs_Mixin, NPV_Mixin
-from pyomo.core.expr.expr_errors import TemplateExpressionError
 from pyomo.core.expr.logical_expr import BooleanExpression
 from pyomo.core.expr.numeric_expr import (
     NumericExpression,
